% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties_csv}
\alias{counties_csv}
\title{Counties CSV}
\format{
A CSV file with 8 columns:
\describe{
\item{""}{Index No.}
\item{"county"}{County name}
\item{"state"}{State}
\item{"geoid"}{Geo ID}
\item{"lat"}{Latitude}
\item{"lon"}{Longitude}
\item{"area_m2"}{Area in square meters}
\item{"jitter_radius"}{Jitter radius in meters}
}
}
\usage{
counties_csv
}
\description{
A CSV dataset containing the name, state, geoid, latitude, longitude, area (m^2), and jitter radius of each county in the United States
}
\keyword{datasets}
