% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties_json}
\alias{counties_json}
\title{Counties JSON}
\format{
A JSON file with 7 fields:
\describe{
\item{"county"}{County name}
\item{"state"}{State}
\item{"geoid"}{Geo ID}
\item{"lat"}{Latitude}
\item{"lon"}{Longitude}
\item{"area_m2"}{Area in square meters}
\item{"jitter_radius"}{Jitter radius in meters}
}
}
\usage{
counties_json
}
\description{
A JSON dataset containing the name, state, geoid, latitude, longitude, area (m^2), and jitter radius of each county in the United States
}
\keyword{datasets}
