% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow}
\alias{flow}
\title{Implement inflow and outflow}
\usage{
flow(loc, week, taxa, n, direction = "forward", save_local = FALSE)
}
\arguments{
\item{loc}{One or more locations as a scalar character each location should
be latitude and longitude separated by a comma, multiple locations are separated by
a semicolon  e.g. "42,-72"   or "42,-72;43.4,-72.7"}

\item{week}{The week number to start at.}

\item{taxa}{a taxa.  Should be either "total" (sum across all species) or
one of the species listed in config/taxa.json}

\item{n}{The number of weeks to project forward (note output will include
the initial week so will have n + 1 images)}

\item{direction}{Which direction to project in "backward" for inflow or "forward" for outflow"}

\item{save_local}{True/False boolean flag - determines whether to save results locally (default: value in config/save_local.flag, or FALSE if not present). If FALSE, results are uploaded to S3 if configured.}
}
\value{
A list with components:

\code{start} a list with:
\code{week}  as input
\code{taxa}, as input
\code{loc},  as input
\code{type} - "inflow" or "outflow"
\code{status} either: "success", "error", "outside mask"
\code{result} a list of information about the images each item includes
\code{week}
\code{url}
\code{legend}
\code{type}
}
\description{
This function is the heart of the inflow and outflow api and does all
the work. It is wrapped by both of those endpoints
}
