% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_colors}
\alias{flow_colors}
\title{Flow colors}
\format{
\subsection{\code{flow_colors}}{

An integer matrix with 256 rows and 3 columns:
\describe{
\item{red}{Red channel intensity (0–255)}
\item{green}{Green channel intensity (0–255)}
\item{blue}{Blue channel intensity (0–255)}
}
}
}
\source{
\url{https://github.com/UMassCDS/avian-influenza-api-backend}
}
\usage{
flow_colors
}
\description{
An integer matrix with 256 rows and 3 columns
}
\keyword{datasets}
