% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_model_from_url}
\alias{load_model_from_url}
\title{Load an RDS model from a public S3 URL directly into memory}
\usage{
load_model_from_url(model_url)
}
\arguments{
\item{model_url}{URL pointing to the RDS file}
}
\value{
R object from the loaded RDS file
}
\description{
Downloads and loads an RDS model directly into memory from a public S3 URL
}
