% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_rescale.R
\name{range_rescale}
\alias{range_rescale}
\title{Rescale data to range between 0 and 1}
\usage{
range_rescale(x, min_value = NULL, max_value = NULL)
}
\arguments{
\item{x}{A vector of values to be rescaled}

\item{min_value, max_value}{(optional) if supplied these represent the
minimum and maximum possible values across a larger pool of data than
just \code{x} if not supplied they are calcualted from \code{x}}
}
\value{
x rescaled.
}
\description{
Setting \code{max_value} to the overall maximum of a larger pool of data allows
consistent re scaling across multiple data sets.
}
