% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{set_s3_config}
\alias{set_s3_config}
\title{Set S3 configuration explicitly}
\usage{
set_s3_config(
  access_key = NULL,
  secret_key = NULL,
  region = NULL,
  bucket = NULL
)
}
\description{
Allows user to set S3 credentials and bucket at runtime. If not set, environment variables or IAM role will be used.
}
