% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Species populations and labels}
\format{
A data frame with 9 rows and 3 columns:
\describe{
\item{species}{Character code for the species.}
\item{label}{Common name of the species.}
\item{population}{Estimated population size.}
}
}
\source{
\url{https://github.com/UMassCDS/avian-influenza-api-backend}
}
\usage{
species
}
\description{
A dataset containing population estimates for various North American duck species
}
\keyword{datasets}
