% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolize_raster_data.R
\name{symbolize_raster_data}
\alias{symbolize_raster_data}
\title{Symbolize raster data as a three band PNG}
\usage{
symbolize_raster_data(
  tiff = NULL,
  png,
  col_palette,
  overwrite = TRUE,
  max_value = NULL,
  rast = NULL
)
}
\arguments{
\item{tiff}{Input tiff file; optional see rast for alternative}

\item{png}{Output png file path.}

\item{col_palette}{Color palette. Matrix of 256 colors (rows) and three (rgb)
columns indicating colors.}

\item{overwrite}{If \code{TRUE} overwrite pre-existing colors}

\item{max_value}{The maximum possible value to represent in the color
gradient}

\item{rast}{A \code{terra::SpatRaster} object with the data to write}
}
\value{
invisible()
}
\description{
Symbolize raster data as a three band PNG
}
