% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow}
\alias{flow}
\title{Implement inflow and outflow}
\usage{
flow(loc, week, taxa, n, direction = "forward", save_local = SAVE_LOCAL)
}
\description{
This function is the heart of the inflow and outflow api and does all
the work. It is wrapped by both of those endpoints
}
