% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_config.R
\name{set_s3_config}
\alias{set_s3_config}
\title{Set S3 Configuration}
\usage{
set_s3_config(
  access_key = NULL,
  secret_key = NULL,
  region = NULL,
  bucket = NULL,
  log = TRUE,
  log_file_path = "./flow_debug.log",
  local_temp_path = "localtmp"
)
}
\arguments{
\item{access_key}{AWS access key (default: NULL)}

\item{secret_key}{AWS secret key (default: NULL)}

\item{region}{AWS region (default: NULL)}

\item{bucket}{S3 bucket name (default: NULL)}

\item{log}{Enable logging (default: TRUE)}

\item{log_file_path}{Path for log file (default: "./flow_debug.log")}

\item{local_temp_path}{Path for local temp files (default: "localtmp")}
}
\description{
Set S3 Configuration
}
